backend/requirements.txt
Django==3.1
djangorestframework==3.11.1
django-rest-auth==0.9.5
django-cors-headers==3.5.0
gunicorn==20.0.4
django-extensions==3.1.5
django-filter==21.1
